requirements.txt
streamlit==1.33.0
pandas==1.5.3
matplotlib==3.6.3
seaborn==0.11.2
python-dotenv==1.0.1
bcrypt==4.2.1
SQLAlchemy==2.0.29
python-dateutil==2.8.2